修改 install.sh

+++ resource/deploy/20211215/install.sh
@@ -124,7 +124,11 @@ fi
 
 echo_info "------ 检查初始配置文件 ------ \n"
 cp "licence.key" "${config_dir}"
-cp "configs.py" "${config_dir}"
+
+if [ ! -f "${datarc_dir}/configs.py" ]; then
+    cp "configs.py" "${config_dir}"
+fi
+
 config_files=("update.sh" "docker-compose.yml" ".env" "configs.py" "licence.key")
 for config_file in "${config_files[@]}"; do
   if [ ! -f "${config_dir}/${config_file}" ]; then
